feat: add notifyRewardAmount function to StakingReward

+++ src/tokens/MockERC20.sol
@@ -3,19 +3,13 @@ pragma solidity ^0.8.26;
 
 import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
-contract MockERC20 is ERC20{
-    
-
-    constructor (string memory name, string memory symbol) ERC20(name,symbol){
-
-        _mint(msg.sender, 10000*10**decimals());
+contract MockERC20 is ERC20 {
+    constructor(string memory name, string memory symbol) ERC20(name, symbol) {
+        _mint(msg.sender, 10000 * 10 ** decimals());
     }
 
-    function mint(address to, uint amount) public{
-        require(amount > 0,"amount must greater than zero");
+    function mint(address to, uint256 amount) public {
+        require(amount > 0, "amount must greater than zero");
         _mint(to, amount);
     }
-
-
-
 }
